refactor(blorktools): rig subpackage of scene
