+++ packages/blorktools/src/asset_debugger/loading-splash/loading-splash.js
@@ -1,6 +1,6 @@
 // loading-splash.js - Updated for overlay integration
 
-import { getCaller } from "../util/log-util";
+import { getCaller } from "../util/state/log-util";
 
 /**
  * Shows the loading splash screen overlay
