+++ packages/blorktools/src/asset_debugger/util/animation/render/image2texture-prerender.js
@@ -22,7 +22,7 @@ import {
 } from "../../state/animation-state";
 import { showStatus } from '../../../modals/html-editor-modal/html-editor-modal';
 import { createMeshInfoPanel } from '../../../modals/html-editor-modal/mesh-info-panel-util';
-import { logAnimationAnalysisReport } from '../../log-util';
+import { logAnimationAnalysisReport } from '../../state/log-util';
 import { startPlayback, updateMeshTexture } from '../playback/animation-playback-util';
 import { createTextureFromIframe } from './iframe2texture-render';
